requirement.txt
asgiref==3.5.2
attrs==22.1.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
lxml==4.9.1
Markdown==3.4.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.2.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
zeep==4.1.0